Move deny_floating_point to wasmi
